atomic_inc_dec: create a new inc_dec table
What I was finding was that I was creating a lot of separate little
function hooks where the only difference was the param number and the
key.  I wanted to support of_node_get() which changes param=0 and
key="$->kobj.kref.refcount.ref.counter".  I would have had to create a
special function and also add it to the list of primitive_funcs[].  Which
seemed like a lot of work, and there are doubtless going to be lots of
other similar functions as well.

So what I did is I borrowed the locking_table idea from check_locking.c.
This table is slightly better though because 1) It has the key.  2) It has
and implied flag instead of the enum.  3)  It has a callback field so you
can specify your own callback.  These features makes it much more flexible.

Also I noticed that in db/fixup_kernel.sh, it had a list of stuff to delete
because it was handled by the test itself.  This was already handled by the
primitive_funcs[] code so it wasn't required any more.  And I suspect that
if it ever worked correctly then the primitive_funcs code wouldn't have
been required.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>