xtensa: fix configuration of endianness
Since gcc 3.4.0 there is no option to specify the endianness for
the Xtensa architecture, so the kernel relies on autodetecting
the endianness and then defining the macros __XTENSA_E{B,L}__.

But this means that sparse's 'arch_big_endian' can't be used
for the predefine.

So, do not predefine these macros anymore, they will transparently
be set directly from the command line.

Reported-by: Peter Zijlstra <peterz@infradead.org>
Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>