fix stream_prev() for invalid (negative) stream
Now that the parent stream is stored as a position, the validity
of a stream can't anymore be tested by checking if its number
is non-negative because inside a struct position stream number
are stored as an unsigned (and changing it to signed would halve
the maximum number of stream).

So, add a check against input_stream_nr before returning the
previous stream.

Fixes: 4c6cbe557c48205f9b3d2aae4c166cd66446b240
Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>