sindex: allow indexing outside the project tree
One possible way to compile the linux kernel is by using the O=<DIR>
parameter to place all generated files outside the source tree.

Prior to this patch, sindex filters processed sources to exclude system
files. The base directory of the project was the current directory.

When compiled outside of the source tree, this may not be the case.
This patch adds a parameter and an environment variable to specify
the source tree.

You can use it like this:

$ make O=$PWD-build C=2 CHECK="sindex -B $PWD add --"

This parameter is also needed for searching if you want to display
the source code line because sindex does not store lines in the database
but reads them from source files.

Signed-off-by: Alexey Gladkov <gladkov.alexey@gmail.com>
Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>