dereferences_param: mark more parameters as dereferenced
This does a few things.  If we call frob(&foo->bar); and frob dereferences
the parameter then "foo" is dereferenced.  Also track assignments.

The other thing that it does is it ignores impossible paths.  Like say
there is a NULL check:

	if (!foo)
		return NULL;
	return foo->bar;

And we always pass a non-NULL then mark the function as dereferencing
"foo".  This feels sort of ugly to me and it is.  But it's also not wrong.

This change does two things.  First it obviously catches more bugs.  But
the other reason to do this is because if a function dereferences a
parameter then we know that the parameter can't be NULL.  This let's us
silence some false positives later about inconsistent NULL checking or
whatever...

Smatch has for many years tried to ignore crappy inconsistent NULL checking
where it was known to be harmless.  In the kernel, we generally frown on
crappy code but sometimes people argue that crappy code is important for
consistency.  It's not a debate I want to be a part of, so try your hardest
to figure out which bugs are real etc.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>