db: don't store unreachable calls
Sometimes people will "temporarily" add a return to the front of a
buggy function:

static int my_function(void)
{
	return 0;  // FIXME

	frob(a); frob(b); frob(c);
}

When that happens then Smatch thinks we aren't holding any locks when we
call frob() and all the variables that we pass to frob() are unknown.  This
is misleading.  When the call is unreachable, then it's better to not
record anything.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>