fix checking if type is void
Sparse warns if a void function returns a non-void expression.

But the check directly compares the returned type with void_ctype,
without taking in account the presence of a SYM_NODE. In
consequence, sparse issues a few false warnings.

Fix this by using is_void_type() to test the returned type.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>