wstring: add support for examination of string initialization
The examination of a string initializer doesn't know about wide
strings. The only thing needed is if the base type is some kind
of char but for wide chars, this type is the same as 'int' and
an array of ints can't be treated the same as an array of chars.

So, do the detection for wide string initializers as:
1) check that the LHS base type is wchar_ctype
2) check that the RHS is a kind of string expression
   (possibly between braces or parenthesis, recursively).

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>