warning: conditionalize "advancing past deep designator"
The warning "advancing past deep designator" is issued when a
multi-level (deep) designator is followed by a non-designated one.

But it's far from clear what is the value of this warning, what
could be the confusion about this situation and what is special
about these 'deep' designators?

There are hundreds such occurrences in the kernel (394 in the configs
I use for testing) and GCC, clang and sparse have no problems to handle
them correctly. This means that there is also 0 chances that they
will be 'corrected'.

So, add conditionalize this warning with a '-Wpast-deep-designator'
and make it false by default.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>