fix evaluate_ptr_add() when sizeof(offset) != sizeof(pointer)
For a binary op, both sides need to be converted to the resulting
type of the usual conversion. For a compound-assignment (which
is equivalent to a binary op followed by an assignment), the LHS
can't be so converted since its type needs to be preserved for
the assignment, so only the RHS is converted at evaluation and
the type of the RHS is used at linearization to convert the LHS.

However, in the case of pointer arithmetics, a number of shortcuts
are taken and as a result additions with mixed sizes can be produced
producing invalid IR.

So, fix this by converting the RHS to the same size as pointers,
as done for 'normal' binops.

Note: On 32-bit kernel, this patch also removes a few warnings
      about non size-preserving casts. It's fine as these warnings
      were designed for when an address would be stored in an
      integer, not for storing an offset like it's the case here.

Reported-by: Valentin Schneider <valentin.schneider@arm.com>
Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>