capped: hard max does not mean capped
The definition of capped means that it's capped to an unknown value so
a hard max does not mean it's capped because that's a known value.  The
reason this is important is because of how this interacts with user data.
Say we test a value against INT_MAX, that's not a valid cap and not going
to be a valid array bounds in real life.  But in the original code we
passed to caller_info that the value is capped.  This silences real
warnings.  But the issue that I cared about is then we have another
function call which does:

	if (!valid_idx(idx)) {

That function is the real cap, but it's not recorded in the database
because it says that the data we passed always starts as capped.

Meanwhile smatch_kernel_user_data.c doesn't count the INT_MAX limit as a
real cap and it was expecting that valid_idx() would set CAPPED_DATA as
a return_state.  So this bug leads to false positives.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>