locking: turn off locking check for non-SMP configs
On uniprocessor configs then there is no spin_lock functions but there are
spin_unlock functions so you end up with tons of double unlock warnings.
That's perhaps an over simplification of the situation and we could check
for mutexes or whatever...  But really uniprocessor configs are not common
and this change will silence a lot of false positives without losing any
real bugs.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>