db/kernel.insert: Insert some vmalloc() limits
Generally Smatch assumes that vmalloc() can only allocate 128MB.  I
doubt that this is the actual limit everywhere, but it's important to
have a limit.  Because if vmalloc() succeeds then we know that the size
is non-zero and we know that it's not ULONG_MAX and this information is
useful to avoid warning about potential integer overflows for example.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>