db/kernel.insert: the device_add() function sets the dev->driver pointer
Smatch has a hard time parsing __device_attach_driver() for a couple
reasons.  It's a weird call back thing.  I guess for asynchronous probing
and Smatch is notoriously bad at understanding threads.  The other reason
is that the device pointer is cast to NULL.

Anyway, this information is important because otherwise Smatch thinks that
the dev->device pointer is NULL.  Then when Smatch tries to parse code
which does:

	client = i2c_new_client_device(adap, info);
	if (i2c_client_has_driver(client)) {
		...

Then Smatch assumes that it's impossible for the client to have a driver
so it gets very confused.  Which leads to false positives.  In this case
it is a lot of dereferencing error pointer false positives.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>