function_hooks: check the return from the implied_return_hooks
The implied_return_hooks are for when we want to set the return value.
Sometimes we have enough data to say what the return is and it returns
one but it if we don't want to set the return it returns zero.  It should
use these returns instead of assuming that the functions always return
true/one.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>