return_to_param: Fix returned function parsing
When we are trying to preserve information about the return value it is
tricky because there isn't always a variable to attach that to.  So what
this code did was if you had a function assignment:

	foo = frob();

Then it stored the return information in "foo".  But if you didn't have a
function assignment:

	frob();

Then it generated an sm->name and symbol based on the call expression.
Which is fine...  But then when I started generating fake assignments,
the state was attached to the fake variable.  So then, when we look it
up, we think that it's not an assignment so the state should be attached
to call expression.  But it's not, we have to look for the fake assignment
and use that instead.

This affects code returns a pointer like:

static inline void *pci_get_drvdata(struct pci_dev *pdev)
{
        return dev_get_drvdata(&pdev->dev);
}

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>