unwind: re-write the check
TLDR: Just add new functions to the table at the start of the file.

This started as a copy and paste of the check_atomic_inc_dec.c code and
builds on it.

The original check_unwind.c code was kind of ancient and it set the state
when the request_irq() function failed.  That's unnecessary.  I removed
the release_function_indicator() calls as well and replaced it with DB
code.

The original code had a bunch of custom written functions to tie the
function call to a variable but with the table, we can just specify a
param/key pair instead.  This is a lot more flexible.  It's simpler to
add new alloc/free pairs as well.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>