db: split ranged implication hooks
The problem here was that misc_register() wasn't getting split into
success/fail paths, but the check_unwind.c wanted to set the state for
the true path.  This was a problem for some of the locking functions
as well.

What I've done is said that if a function has return implication hard
coded then just split it as much as possible.  This is perhaps wasteful
and also maybe 100% correct because maybe there are some places where
the DB code doesn't know how to split it into success/fail so we'd have
to do that manually...  It works for now.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>