db: fix db_incomplete
There were two problems with this.  First the db/fixup_all.sh deletes a
bunch of returns.  That code is pretty old, and I question whether it's
still required.  I am going to investigate.

Second the kernel handles __memset() with and alias:

extern void *__memset(void *s, int c, size_t n) __alias(memset);

I'm not sure how to handle this.  But what I can do is only mark the DB
incomplete if there are defined functions where we don't get results.
This is basically functions in the same file that we are parsing.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>