param_key: preserve parameter assignments
The original code used get_other_name_sym(), the problem with that was if
you had code like:

void free_something(struct device *dev)
{
	struct my_data *p = dev->data;

	dev->data = NULL;
	free(p->xxx);
}

The code frees $->data->xxx but in the "dev->data = NULL;" assignment
clears that relationship so we can't use smatch_assignment.c.  This is a
smaller replacement for that.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>