unwind: handle freeing in other functions better
Sometimes we have a free function that frees $->foo->bar but in the caller
we can't tie $->foo->bar to something that we allocated.  This patch takes
a hard line to try silence false positives in that situation.

If a function hasn't allocated anything then it doesn't change anything.
I've added the "fn_has_alloc" flag to indicate that.

But if it has allocated something then we take the "bar" from "$->foo->bar"
and mark everything that ends in "bar" as freed.

This code works pretty well.  The remaining issue that I see is that there
are cases where Smatch still has trouble connecting a freed pointer to a
parameter.  For example, netdev_priv() is tricky.  The other issue is that
Smatch doesn't understand kobj release functions.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>