math: improve AND handling
I wrote this code some time ago and have been running it, so presumably it
is good code.  The original code seems to have assumed that the smallest
value was on the left.

We want to get the minimum maximum.  That's the highest shared bit.

This fixes parsing of array_index_nospec().

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>