capped: change how binops are handled
The original code was a nonsense mess.  The rules for capped are not
that complicated.  There are some binops which cap things like MOD and
bitwise AND.  For the rest, if we have both sides capped then the result
is capped.  Or if one side is capped and the other side is known then the
result is capped.  But if neither side is capped (even if the result is
known) then the result is uncapped.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>