flow: fake a parameter assignment for literals
What I want is that there should be just one callback to handle inserting
data into caller_info and return_info.  That's the point of the
fake assignments.

So say we want to record that parameter value then we could just use
the extra state.  But if we don't fake a parameter assign for literals
then there is no state and we have to use a separate callback.  So fake
the literal assignments.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>