fix Hurd PATH_MAX ...
The fix quickly added for Hurd not defining PATH_MAX is not correct:
the local define must be guaranteed to be done only after <limits.h>
is included, which is not.

Since this problem only exists with Hurd, simply conditionalize
the local define by __gnu_hurd__. Horrible but well ... Hurd.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>