fix evaluation of pointer to bool conversions
The pointer to bool conversion used an indirect intermediate
conversion to an int because the pointer was compared to 0
and not to a null pointer. The final result is the same
but the intermediate conversion generated an unneeded OP_PTRTOU
instruction which made some tests to fail.

Fix this by directly comparing to a null pointer of the same
type as the type to convert.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>