builtin: make eval_sync_compare_and_swap() more generic
Most __sync_* or __atomic_* builtin functions have one or
more arguments having its real type determined by the first
argument: either the same type (a pointer to an integral type)
or the type of the object it points to.

Currently, only __sync_{bool,val}_compare_and_swap() are handled
but lots of very similar variants would be needed for the others.
So, make it a generic function, able to handle all these builtins.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>