math: store all constant EXPR_BINOP results
Smatch was caching the results when the whole big expression was constant
but that didn't work for huge expressions where only half of it was
known.

The fix for this is that there is actually unused space in the expression
struct where we can store any constant results.  In practice the hugest
expressions are always BINOPs so this is a good speed up.  Probably I can
remove the cache entirely now.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>