smatch: save less stuff, run faster
The story behind this patch was that I was trying to speed up Smatch so
I looked through the cur_slist and tried to spot useless information.
Sometimes it's just fake stuff and that's easy to ignore.

In real_absolute it was looking in the DB, and then moving that information
to the slist which is useless.  Just leave it in the DB unless absolutely
required.

In check_leaks() it was calling get_implied_value() which queries the DB.
Try to avoid accessing the DB until it's required.  There is probably a
less here, and possibly get_implied_value() should not even access the DB
at all?

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>