param_limit/filter/set: use param_key code
The param_key code is intended to be the standard, central way to map a
variable to a param/key pair.  It's more powerful than the open coded
mappings that this code use to use.

The thing of it as well, is that these checks need to work in sync
otherwise it leads to weird bugs.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>