flow: don't parse inlines again if they take a long time
I was trying to speed up Smatch and so I made this change which records
how long it takes to parse a function.  Then when we're looking at inlines
don't parse it again if the inline took a long time to parse.

I don't know if this code is actually worthwhile but it felt like a good
idea.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>