kill dead instructions before any other simplifications
Dead instructions are removed when simplify_instruction() is
called but this is done in various places, depending on the kind
of instructions, sometimes after other simplifications.

Change this by using the instruction's flag OPF_TARGET at the very
beginning of simplify_instruction() to test which instructions are
dead and thus can be removed.

Note: OP_CALLs are special here as they're considered as always
      returning a value but only calls to pure functions are
      removed. This is OK since pure functions should always
      return a value.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>