fix testing if a OP_CALL's function is pure
kill_instruction() will kill an OP_CALL but only if it's a
forced kill or if the corresponding function is pure.

However, only functions called via a symbol pseudo are so killed.
Those called via a function pointer are not because only symbol
pseudos contain the function type needed to test the presence of
the MOD_PURE modifier.

Fix this by using the function type always available in the
instruction's ::fntypes member.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>