cmp: add signed/unsigned to opcode table
The opcode table allows to efficiently store some properties of
the IR instructions and the correspondence between some of them.

One of these correspondences the 'signed' / 'unsigned' version
of otherwise identical instructions. This is useful for some
transformation of compare instructions but is not present yet
in the table. So, add this now.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>