testsuite: add a new tag: check-output-match
The current tags check-output-contains/excludes/pattern are
quite powerful, universal, but they often need 'complex' regular
expressions with escaping which make them not so nice to read.

For testing IR results, a very common pattern is:
	this instruction must have this (kind of) operand.

So, make a new tag for this. It does nothing than can't be done
with done with the previous ones, on  the contrary, but is much
simpler to use:
	check-output-match(instruction): operand

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>