testsuite: add a new tag: check-output-returns
The current tags check-output-contains/excludes/pattern are
quite powerful and the new check-output-match is easy to use
but it can be even simpler. Indeed, a lot of IR simplifications/
canonicalizations can be tested by checking that the expression
to be tested is equivalent to another one. This is less precise
than some more specific tests but:
* it's a big advantage because it's less sensitive to 'noise'
  like the exact number used by the pseudos or to the results
  of some new simplifications or canonicalizations
* very often, this equivalence is what really matters and not
  the exact transformation.

So, add a new utra-simple-to-use tag: just ask that all functions
of the tests return the same specified value (usually 1):
	check-output-returns: <value>

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>