param_key: fix an ordering bug
The problem is when you have something like "foo = param".  It was
first recording that "foo is now $" (the param key).  Then it was recording
the smatch_extra assignment "$ is now whatever range list".  And that's
clearly wrong because we haven't set the parameter to anything.  Plus that
would be a stack assignment so we can't PARAM_SET it.

This fix for this is slightly weird, but if we get the ordering right then
I believe it just works.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>